Learn to Program with C++ 17 module 9 added

+++ 01_Learn_to_Program_with_C++_17/Module_03_StreamIO_and_Local_Variables/output.cpp
@@ -19,7 +19,7 @@ int main() {
 
     // type safety; be careful.
     // e.g. of data loss
-    i = 4.9;
+    i = static_cast<int>(4.9); // e.g of explicit data type casting
     int j = 9/5;
     cout << "i is " << i << '\n' << "j is " << j << '\n';
